Pass instance host-id to Quantum using port bindings extension.
This is a nova change to pass the hostname on which the instance was started to
Quantum. The assoociated quantum OVSplugin changes pass this hostname to
hardware device vendors management stations to allow them to segment the network
in a more precise manner (for example automatically trunk the vlan on the
physical switch port connected to the compute host on which the VM instance was
started).

Implements: blueprint vm-host-quantum

Change-Id: I510be53aeda0608a49378ca8eb4766de0af590f3
Signed-off-by: Gary Kotton <gkotton@redhat.com>
Signed-off-by: Aaron Rosen <arosen@nicira.com>